started moving everything to separate methods
